Function 1:F1
Reads the unit name and students' marks from a given text file. The user will provide the file name. The file contains the unit name and the list 
of students with their names, student IDs and marks for three assignments. The file also contains lines, which are comments, and your program 
should check to ignore them when reading the students' marks.